Initial commit of puppet network module.
This is a new module to support traditional (ie *not* NetworkManager)
network interface configuration on RedHat- or Fedora-based systems.  It
allows for static, dhcp, and bootp configuration of normal and bonded
interfaces.  There is support for aliases on normal and bonded
interfaces.  It can configure static routes.  It can configure MTU,
ETHTOOL_OPTS, and BONDING_OPTS on a per-interface basis.