More fixes to allow bonding module to autoload.
This complements cf8ca14 and makes sure that BONDING_OPTS= ends up in
the ifcfg-bondX file and not in ifcfg-ethX even if the user does not
specify the $bonding_opts paramter.  This allows the ifup scripts to
autoload the kernel bonding module.

Also make sure that bond slaves are ifup'ed during their creation so
that the master will start.