network::if::static:  Add VLAN capability
Adds missing VLAN capability to network::if::static.

The existing documentation for using VLAN:s on static interfaces says to
instantiate network::global with the parameter vlan set to yes, such as:
    class{'network::global':
      vlan => 'yes'
    }

This isn't sufficient. ifcfg-$ifname requires also the parameter
'VLAN=yes' defined, for a VLAN interface.

With this approach, that parameter is explicitly defined to
network::static::if, such as:

    'network::static::if' { 'eth0.330':
      ensure    => 'up',
      vlan      => true,
      ipaddress => '10.2.3.248',
      netmask   => '255.255.255.0',
    }

This makes it abundandly clear what's going on.